Put template into its own header file, and ...
add 1d interpolation interface.
